added confirmation before deleting the product

+++ admin/view_products.php
@@ -239,14 +239,14 @@ color:white;
 {
 ?>
   <tr>
-  <td><img src="<?php echo $row['image']?>" alt="Product Image"></td>
+  <td><img src="../photos/<?php echo $row['image'] ?>" alt="Product Image"></td>
 <td><?php echo $row['name']?></td>
 <td class="description"><?php echo $row['description']?></td>
 <td><?php echo $row['price']?></td>
 <td><?php echo $row['quantity']?></td>
 <td>
 <button class="btn edit">Edit</button>
-<a href="deleteproduct.php?id=<?php echo $row['id']; ?>" class="btn delete">Delete</a>
+<a onclick="return confirm('Are you sure you want to delete this product ?')" href="deleteproduct.php?id=<?php echo $row['id']; ?>" class="btn delete">Delete</a>
 </td>
 </tr>
 <?php } ?>
